refactor: @PreviewParameterProvider 적용
각 화면의 component 파일별로 분리 -> preview 적용
