+++ core/designsystem/src/main/kotlin/com/unifest/android/core/designsystem/component/LoadingWheel.kt
@@ -24,7 +24,7 @@ fun LoadingWheel(
 
 @ComponentPreview
 @Composable
-fun LoadingWheelPreview() {
+private fun LoadingWheelPreview() {
     UnifestTheme {
         LoadingWheel()
     }
