+++ core/designsystem/src/main/kotlin/com/unifest/android/core/designsystem/component/NetworkImage.kt
@@ -49,7 +49,7 @@ fun NetworkImage(
 
 @ComponentPreview
 @Composable
-fun NetworkImagePreview() {
+private fun NetworkImagePreview() {
     UnifestTheme {
         NetworkImage(
             imgUrl = "",
