+++ core/designsystem/src/main/kotlin/com/unifest/android/core/designsystem/component/OutlinedButton.kt
@@ -46,7 +46,7 @@ fun UnifestOutlinedButton(
 
 @ComponentPreview
 @Composable
-fun UnifestOutlinedButtonPreview() {
+private fun UnifestOutlinedButtonPreview() {
     UnifestTheme {
         UnifestOutlinedButton(
             onClick = {},
