+++ core/ui/src/main/kotlin/com/unifest/android/core/ui/component/BoothFilterChips.kt
@@ -38,7 +38,7 @@ fun BoothFilterChips(
 
 @ComponentPreview
 @Composable
-fun BoothFilterChipsPreview() {
+private fun BoothFilterChipsPreview() {
     UnifestTheme {
         BoothFilterChips(
             onChipClick = {},
