+++ core/ui/src/main/kotlin/com/unifest/android/core/ui/component/EmptyLikedBoothItem.kt
@@ -52,7 +52,7 @@ fun EmptyLikedBoothItem(
 
 @ComponentPreview
 @Composable
-fun EmptyLikedBoothItemPreview() {
+private fun EmptyLikedBoothItemPreview() {
     UnifestTheme {
         EmptyLikedBoothItem(
             modifier = Modifier
