+++ feature/festival/src/main/kotlin/com/unifest/android/feature/festival/FestivalSearchResults.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import com.unifest.android.core.common.extension.noRippleClickable
 import com.unifest.android.core.common.utils.formatToString
@@ -40,6 +41,9 @@ import com.unifest.android.core.designsystem.theme.Content6
 import com.unifest.android.core.model.FestivalModel
 import com.unifest.android.feature.festival.viewmodel.FestivalUiAction
 import com.unifest.android.core.designsystem.R
+import com.unifest.android.core.designsystem.theme.UnifestTheme
+import com.unifest.android.feature.festival.preview.FestivalPreviewParameterProvider
+import com.unifest.android.feature.festival.viewmodel.FestivalUiState
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 
@@ -191,70 +195,14 @@ fun FestivalSearchResultItem(
 
 @ComponentPreview
 @Composable
-fun FestivalSearchResultsPreview() {
-    FestivalSearchResults(
-        searchResults = persistentListOf(
-            FestivalModel(
-                1,
-                1,
-                "https://picsum.photos/36",
-                "서울대학교",
-                "서울",
-                "설대축제",
-                "2024-04-21",
-                "2024-04-23",
-                126.957f,
-                37.460f,
-            ),
-            FestivalModel(
-                2,
-                2,
-                "https://picsum.photos/36",
-                "연세대학교",
-                "서울",
-                "연대축제",
-                "2024-04-21",
-                "2024-04-23",
-                126.957f,
-                37.460f,
-            ),
-            FestivalModel(
-                3,
-                3,
-                "https://picsum.photos/36",
-                "고려대학교",
-                "서울",
-                "고대축제",
-                "2024-04-21",
-                "2024-04-23",
-                126.957f,
-                37.460f,
-            ),
-            FestivalModel(
-                4,
-                4,
-                "https://picsum.photos/36",
-                "성균관대학교",
-                "서울",
-                "성대축제",
-                "2024-04-21",
-                "2024-04-23",
-                126.957f,
-                37.460f,
-            ),
-            FestivalModel(
-                5,
-                5,
-                "https://picsum.photos/36",
-                "건국대학교",
-                "서울",
-                "건대축제",
-                "2024-04-21",
-                "2024-04-23",
-                126.957f,
-                37.460f,
-            ),
-        ),
-        onFestivalUiAction = {},
-    )
+fun FestivalSearchResultsPreview(
+    @PreviewParameter(FestivalPreviewParameterProvider::class)
+    festivalUiState: FestivalUiState,
+) {
+    UnifestTheme {
+        FestivalSearchResults(
+            searchResults = festivalUiState.likedFestivals,
+            onFestivalUiAction = {},
+        )
+    }
 }
