+++ feature/home/src/main/kotlin/com/unifest/android/feature/home/component/Calendar.kt
@@ -1,4 +1,4 @@
-package com.unifest.android.feature.home
+package com.unifest.android.feature.home.component
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
@@ -67,6 +67,9 @@ import com.unifest.android.core.designsystem.theme.LightRed
 import com.unifest.android.core.designsystem.theme.Title5
 import com.unifest.android.core.designsystem.theme.UnifestTheme
 import com.unifest.android.core.model.FestivalModel
+import com.unifest.android.feature.home.clickable
+import com.unifest.android.feature.home.displayText
+import com.unifest.android.feature.home.rememberFirstVisibleMonthAfterScroll
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
