+++ feature/liked-booth/src/main/kotlin/com/unifest/android/feature/liked_booth/viewmodel/LikedBoothUiState.kt
@@ -7,7 +7,6 @@ import kotlinx.collections.immutable.persistentListOf
 data class LikedBoothUiState(
     val isLoading: Boolean = false,
     val likedBooths: ImmutableList<LikedBoothModel> = persistentListOf(),
-    // val likedBoothList: ImmutableList<BoothDetailModel> = persistentListOf(),
     val isServerErrorDialogVisible: Boolean = false,
     val isNetworkErrorDialogVisible: Boolean = false,
 )
