+++ feature/main/src/main/kotlin/com/unifest/android/feature/main/component/MainBottomBar.kt
@@ -1,0 +1,131 @@
+package com.unifest.android.feature.main.component
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.slideIn
+import androidx.compose.animation.slideOut
+import androidx.compose.foundation.background
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.selection.selectable
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.IntOffset
+import androidx.compose.ui.unit.dp
+import com.unifest.android.core.designsystem.ComponentPreview
+import com.unifest.android.core.designsystem.theme.BottomMenuBar
+import com.unifest.android.core.designsystem.theme.UnifestTheme
+import com.unifest.android.feature.main.MainTab
+import kotlinx.collections.immutable.ImmutableList
+import kotlinx.collections.immutable.toImmutableList
+
+@Composable
+internal fun MainBottomBar(
+    visible: Boolean,
+    tabs: ImmutableList<MainTab>,
+    currentTab: MainTab?,
+    onTabSelected: (MainTab) -> Unit,
+) {
+    AnimatedVisibility(
+        visible = visible,
+        enter = fadeIn() + slideIn { IntOffset(0, it.height) },
+        exit = fadeOut() + slideOut { IntOffset(0, it.height) },
+    ) {
+        Box(modifier = Modifier.background(MaterialTheme.colorScheme.surface)) {
+            Column {
+                HorizontalDivider(color = MaterialTheme.colorScheme.outline)
+                Row(
+                    modifier = Modifier
+                        .navigationBarsPadding()
+                        .fillMaxWidth()
+                        .height(64.dp),
+                    verticalAlignment = Alignment.CenterVertically,
+                ) {
+                    tabs.forEach { tab ->
+                        MainBottomBarItem(
+                            tab = tab,
+                            selected = tab == currentTab,
+                            onClick = {
+                                if (tab != currentTab) {
+                                    onTabSelected(tab)
+                                }
+                            },
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun RowScope.MainBottomBarItem(
+    tab: MainTab,
+    selected: Boolean,
+    onClick: () -> Unit,
+) {
+    Box(
+        modifier = Modifier
+            .weight(1f)
+            .fillMaxHeight()
+            .selectable(
+                selected = selected,
+                indication = null,
+                role = null,
+                interactionSource = remember { MutableInteractionSource() },
+                onClick = onClick,
+            ),
+        contentAlignment = Alignment.Center,
+    ) {
+        Column(
+            horizontalAlignment = Alignment.CenterHorizontally,
+        ) {
+            Icon(
+                imageVector = if (selected) ImageVector.vectorResource(tab.selectedIconResId)
+                else ImageVector.vectorResource(tab.iconResId),
+                contentDescription = tab.contentDescription,
+                tint = Color.Unspecified,
+            )
+            Spacer(modifier = Modifier.height(5.dp))
+            Text(
+                text = tab.label,
+                color = if (selected) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.onSecondary,
+                fontWeight = if (selected) FontWeight.SemiBold
+                else FontWeight.Normal,
+                style = BottomMenuBar,
+            )
+        }
+    }
+}
+
+@ComponentPreview
+@Composable
+private fun MainBottomBarPreview() {
+    UnifestTheme {
+        MainBottomBar(
+            visible = true,
+            tabs = MainTab.entries.toImmutableList(),
+            currentTab = MainTab.HOME,
+            onTabSelected = {},
+        )
+    }
+}
