+++ feature/menu/src/main/kotlin/com/unifest/android/feature/menu/MenuScreen.kt
@@ -4,7 +4,6 @@ import android.content.pm.PackageManager
 import android.widget.Toast
 import androidx.compose.animation.core.LinearOutSlowInEasing
 import androidx.compose.animation.core.tween
-import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -17,16 +16,12 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
@@ -36,18 +31,15 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalUriHandler
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
@@ -55,19 +47,13 @@ import com.unifest.android.core.common.ObserveAsEvents
 import com.unifest.android.core.common.UiText
 import com.unifest.android.core.designsystem.R
 import com.unifest.android.core.designsystem.component.NetworkErrorDialog
-import com.unifest.android.core.designsystem.component.NetworkImage
 import com.unifest.android.core.designsystem.component.ServerErrorDialog
 import com.unifest.android.core.designsystem.component.TopAppBarNavigationType
 import com.unifest.android.core.designsystem.component.UnifestHorizontalDivider
 import com.unifest.android.core.designsystem.component.UnifestTopAppBar
-import com.unifest.android.core.designsystem.theme.Content6
 import com.unifest.android.core.designsystem.theme.Content7
-import com.unifest.android.core.designsystem.theme.Content8
-import com.unifest.android.core.designsystem.theme.MenuTitle
 import com.unifest.android.core.designsystem.theme.Title3
 import com.unifest.android.core.designsystem.theme.UnifestTheme
-import com.unifest.android.core.model.FestivalModel
-import com.unifest.android.core.model.LikedBoothModel
 import com.unifest.android.core.ui.DevicePreview
 import com.unifest.android.core.ui.component.EmptyLikedBoothItem
 import com.unifest.android.core.ui.component.LikedBoothItem
@@ -76,12 +62,14 @@ import com.unifest.android.feature.festival.viewmodel.FestivalUiAction
 import com.unifest.android.feature.festival.viewmodel.FestivalUiEvent
 import com.unifest.android.feature.festival.viewmodel.FestivalUiState
 import com.unifest.android.feature.festival.viewmodel.FestivalViewModel
+import com.unifest.android.feature.menu.component.FestivalItem
+import com.unifest.android.feature.menu.component.MenuItem
+import com.unifest.android.feature.menu.preview.MenuPreviewParameterProvider
 import com.unifest.android.feature.menu.viewmodel.ErrorType
 import com.unifest.android.feature.menu.viewmodel.MenuUiAction
 import com.unifest.android.feature.menu.viewmodel.MenuUiEvent
 import com.unifest.android.feature.menu.viewmodel.MenuUiState
 import com.unifest.android.feature.menu.viewmodel.MenuViewModel
-import kotlinx.collections.immutable.persistentListOf
 import timber.log.Timber
 
 @Composable
@@ -141,7 +129,6 @@ internal fun MenuRoute(
     )
 }
 
-@OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun MenuScreen(
     padding: PaddingValues,
@@ -260,21 +247,23 @@ fun MenuScreen(
                         items = menuUiState.likedBooths.take(3),
                         key = { _, booth -> booth.id },
                     ) { index, booth ->
+                        Modifier
+                            .clickable {
+                                onMenuUiAction(MenuUiAction.OnLikedBoothItemClick(booth.id))
+                            }
                         LikedBoothItem(
                             booth = booth,
                             index = index,
                             totalCount = menuUiState.likedBooths.size,
                             deleteLikedBooth = { onMenuUiAction(MenuUiAction.OnToggleBookmark(booth)) },
-                            modifier = Modifier
-                                .clickable {
-                                    onMenuUiAction(MenuUiAction.OnLikedBoothItemClick(booth.id))
-                                }
-                                .animateItemPlacement(
-                                    animationSpec = tween(
-                                        durationMillis = 500,
-                                        easing = LinearOutSlowInEasing,
-                                    ),
+                            modifier = Modifier.animateItem(
+                                fadeInSpec = null,
+                                fadeOutSpec = null,
+                                placementSpec = tween(
+                                    durationMillis = 500,
+                                    easing = LinearOutSlowInEasing,
                                 ),
+                            ),
                         )
                     }
                 }
@@ -350,137 +339,16 @@ fun MenuScreen(
     }
 }
 
-@Composable
-fun FestivalItem(
-    festival: FestivalModel,
-    onMenuUiAction: (MenuUiAction) -> Unit,
-) {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-        modifier = Modifier
-            .padding(vertical = 10.dp)
-            .clickable {
-                onMenuUiAction(MenuUiAction.OnLikedFestivalItemClick(festival.schoolName))
-            },
-    ) {
-        Box(
-            modifier = Modifier
-                .size(65.dp)
-                .shadow(
-                    elevation = 6.dp,
-                    shape = CircleShape,
-                )
-                .background(MaterialTheme.colorScheme.tertiaryContainer, CircleShape)
-                .padding(5.dp),
-            contentAlignment = Alignment.Center,
-        ) {
-            NetworkImage(
-                imgUrl = festival.thumbnail,
-                contentDescription = "Festival Thumbnail",
-                modifier = Modifier
-                    .size(60.dp)
-                    .clip(CircleShape),
-                placeholder = painterResource(id = R.drawable.item_placeholder),
-            )
-        }
-        Spacer(modifier = Modifier.height(8.dp))
-        Text(
-            text = festival.schoolName,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            style = Content6,
-            textAlign = TextAlign.Center,
-        )
-        Spacer(modifier = Modifier.height(2.dp))
-        Text(
-            text = festival.festivalName,
-            color = MaterialTheme.colorScheme.onBackground,
-            style = MenuTitle,
-            textAlign = TextAlign.Center,
-        )
-    }
-}
-
-@Composable
-fun MenuItem(
-    icon: ImageVector,
-    title: String,
-    onClick: () -> Unit,
-) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(MaterialTheme.colorScheme.background)
-            .clickable(onClick = onClick)
-            .padding(25.dp),
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = "Menu Icon",
-            tint = Color.Unspecified,
-        )
-        Spacer(modifier = Modifier.width(16.dp))
-        Text(
-            text = title,
-            color = MaterialTheme.colorScheme.onBackground,
-            style = Content8,
-        )
-    }
-}
-
 @DevicePreview
 @Composable
-fun MenuScreenPreview() {
+fun MenuScreenPreview(
+    @PreviewParameter(MenuPreviewParameterProvider::class)
+    menuUiState: MenuUiState,
+) {
     UnifestTheme {
         MenuScreen(
             padding = PaddingValues(),
-            menuUiState = MenuUiState(
-                festivals = persistentListOf(
-                    FestivalModel(
-                        1,
-                        1,
-                        "https://picsum.photos/36",
-                        "서울대학교",
-                        "서울",
-                        "설대축제",
-                        "2024-04-21",
-                        "2024-04-23",
-                        126.957f,
-                        37.460f,
-                    ),
-                    FestivalModel(
-                        2,
-                        2,
-                        "https://picsum.photos/36",
-                        "연세대학교",
-                        "서울",
-                        "연대축제",
-                        "2024-04-21",
-                        "2024-04-23",
-                        126.957f,
-                        37.460f,
-                    ),
-                ),
-                likedBooths = persistentListOf(
-                    LikedBoothModel(
-                        id = 1,
-                        name = "부스 이름",
-                        category = "부스 카테고리",
-                        description = "부스 설명",
-                        location = "부스 위치",
-                        warning = "학과 전용 부스",
-                    ),
-                    LikedBoothModel(
-                        id = 2,
-                        name = "부스 이름",
-                        category = "부스 카테고리",
-                        description = "부스 설명",
-                        location = "부스 위치",
-                        warning = "학과 전용 부스",
-                    ),
-                ),
-            ),
+            menuUiState = menuUiState,
             festivalUiState = FestivalUiState(),
             appVersion = "1.0.0",
             onMenuUiAction = {},
