+++ feature/menu/src/main/kotlin/com/unifest/android/feature/menu/component/MenuItem.kt
@@ -1,0 +1,47 @@
+package com.unifest.android.feature.menu.component
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+import com.unifest.android.core.designsystem.theme.Content8
+
+@Composable
+fun MenuItem(
+    icon: ImageVector,
+    title: String,
+    onClick: () -> Unit,
+) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(MaterialTheme.colorScheme.background)
+            .clickable(onClick = onClick)
+            .padding(25.dp),
+    ) {
+        Icon(
+            imageVector = icon,
+            contentDescription = "Menu Icon",
+            tint = Color.Unspecified,
+        )
+        Spacer(modifier = Modifier.width(16.dp))
+        Text(
+            text = title,
+            color = MaterialTheme.colorScheme.onBackground,
+            style = Content8,
+        )
+    }
+}
